user_manager/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
logfmter==0.0.8
